doc: refactor and modularize examples with shared utilities
- Moved reusable logic (e.g., Anvil setup, token management) into `examples/common`.
- Simplified `mint_position_basic.rs` and `mint_position_permit2.rs` by utilizing shared helper modules.
- Removed `dotenv` and `once_cell` from dependencies; replaced with centralized constants.
- Updated tests to eliminate `test-utils` feature, improving modularity.
- Adjusted `Cargo.toml` to reflect dependency changes and code restructure.
